requirements.txt
flask>=2.3.0
google-genai>=0.8.0
pdfplumber>=0.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
